Updated references to FormAddSubmission to be singular
